// src/app/models/interface/client.model.ts
export interface Client {
  id: number;
  prenom: string;
  nom_societe: string;
  nom: string;
  tel1: string;
  tel2: string;
  whatsapp: string;
  facebook_page: string;
  instagram_account: string;
  linkedin_page: string;
  site_web: string;
  email: string;
  pays_id: number;
  gouvernerat_id: number;
  adresse: string;
  matricul_fiscal: string;
  secteur: string;
  notes: string;
  label_id: number;
  logo: string;
}